+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.4.0] - 2025-12-09
 
+- Added new Aerofly FS4 airports after update
 - Added community airports in separate file `data/airports-custom.md`
 - Improved search in Mapbox map
 - Improved handling of small US airports